+++ src/main/java/org/opentripplanner/routing/algorithm/mapping/RoutingRequestToFilterChainMapper.java
@@ -27,7 +27,7 @@ public class RoutingRequestToFilterChainMapper {
     boolean maxNumberOfItinerariesCropHead,
     Consumer<Itinerary> maxLimitReachedSubscriber,
     boolean wheelchairAccessible,
-    float wheelchairMaxSlope
+    double wheelchairMaxSlope
   ) {
     var builder = new ItineraryListFilterChainBuilder(sortOrder);
